+++ src/main/java/org/maas/Objects/Location.java
@@ -1,4 +1,4 @@
-package org.team_pjt.objects;
+package org.maas.objects;
 
 public class Location {
     private Double fX;
