+++ src/main/java/org/maas/Objects/Order.java
@@ -1,4 +1,4 @@
-package org.team_pjt.objects;
+package org.maas.objects;
 
 import jade.core.AID;
 import org.json.JSONObject;
