+++ src/main/java/org/maas/agents/BaseAgent.java
@@ -1,4 +1,4 @@
-package org.team_pjt.agents;
+package org.maas.agents;
 
 import jade.core.Agent;
 import jade.core.AID;
